// frontend/types/pessoas.ts
export interface Pessoa{
    id?: number
    nome?: string
    nascimento?: string
}

export interface FiltroPessoa{
    Id?: number
    Nome?: string
}

export interface NovaPessoa{
    nome: string
    nascimento: string
}

export interface AlteraPessoa{
    id?: number
    nome?: string
    nascimento?: string
}